Address comments and improve the code structure.
